refactor: organize componentes files
